chore: renamed runtime-node to http-server
