+++ lerna.json
@@ -31,7 +31,7 @@
     "packages/runtime-html",
     "packages/runtime-html-browser",
     "packages/runtime-html-jsdom",
-    "packages/runtime-node",
+    "packages/http-server",
     "packages/scheduler",
     "packages/scheduler-dom",
     "packages/scheduler-node",
