+++ packages/http-server/.eslintrc.js
