+++ packages/http-server/README.md
@@ -3,19 +3,19 @@
 [![TypeScript](https://img.shields.io/badge/%3C%2F%3E-TypeScript-%230074c1.svg)](http://www.typescriptlang.org/)
 [![CircleCI](https://circleci.com/gh/aurelia/aurelia.svg?style=shield)](https://circleci.com/gh/aurelia/aurelia)
 [![Test Coverage](https://api.codeclimate.com/v1/badges/5ac0e13689735698073a/test_coverage)](https://codeclimate.com/github/aurelia/aurelia/test_coverage)
-[![npm](https://img.shields.io/npm/v/@aurelia/runtime-node.svg?maxAge=3600)](https://www.npmjs.com/package/@aurelia/runtime-node)
-# @aurelia/runtime-node
+[![npm](https://img.shields.io/npm/v/@aurelia/http-server.svg?maxAge=3600)](https://www.npmjs.com/package/@aurelia/http-server)
+# @aurelia/http-server
 
 ## Installing
 
 For the latest stable version:
 
 ```bash
-npm i @aurelia/runtime-node
+npm i @aurelia/http-server
 ```
 
 For our nightly builds:
 
 ```bash
-npm i @aurelia/runtime-node@dev
+npm i @aurelia/http-server@dev
 ```
