+++ packages/http-server/src/interfaces.ts
